parsecv/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dotenv==0.9.9
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.4
packaging==24.2
pluggy==1.5.0
preshed==3.0.9
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-env==1.1.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
responses==0.25.7
rich==14.0.0
schedule==1.2.2
setuptools==78.1.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
thinc==8.3.6
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.2
